ASN.1: Support wider bit sets (fix #514)

+++ lib/roken/parse_units.h
@@ -51,7 +51,7 @@
 
 struct units {
     const char *name;
-    unsigned mult;
+    unsigned long long mult;
 };
 
 ROKEN_LIB_FUNCTION int ROKEN_LIB_CALL
